refactor repo routes
