+++ packages/api/src/client/index.ts
@@ -35,6 +35,7 @@ import * as ComAtprotoRepoBatchWrite from './types/com/atproto/repo/batchWrite'
 import * as ComAtprotoRepoCreateRecord from './types/com/atproto/repo/createRecord'
 import * as ComAtprotoRepoDeleteRecord from './types/com/atproto/repo/deleteRecord'
 import * as ComAtprotoRepoDescribe from './types/com/atproto/repo/describe'
+import * as ComAtprotoRepoGetBlob from './types/com/atproto/repo/getBlob'
 import * as ComAtprotoRepoGetRecord from './types/com/atproto/repo/getRecord'
 import * as ComAtprotoRepoListRecords from './types/com/atproto/repo/listRecords'
 import * as ComAtprotoRepoPutRecord from './types/com/atproto/repo/putRecord'
@@ -123,6 +124,7 @@ export * as ComAtprotoRepoBatchWrite from './types/com/atproto/repo/batchWrite'
 export * as ComAtprotoRepoCreateRecord from './types/com/atproto/repo/createRecord'
 export * as ComAtprotoRepoDeleteRecord from './types/com/atproto/repo/deleteRecord'
 export * as ComAtprotoRepoDescribe from './types/com/atproto/repo/describe'
+export * as ComAtprotoRepoGetBlob from './types/com/atproto/repo/getBlob'
 export * as ComAtprotoRepoGetRecord from './types/com/atproto/repo/getRecord'
 export * as ComAtprotoRepoListRecords from './types/com/atproto/repo/listRecords'
 export * as ComAtprotoRepoPutRecord from './types/com/atproto/repo/putRecord'
@@ -565,6 +567,17 @@ export class RepoNS {
       .call('com.atproto.repo.describe', params, undefined, opts)
       .catch((e) => {
         throw ComAtprotoRepoDescribe.toKnownErr(e)
+      })
+  }
+
+  getBlob(
+    params?: ComAtprotoRepoGetBlob.QueryParams,
+    opts?: ComAtprotoRepoGetBlob.CallOptions,
+  ): Promise<ComAtprotoRepoGetBlob.Response> {
+    return this._service.xrpc
+      .call('com.atproto.repo.getBlob', params, undefined, opts)
+      .catch((e) => {
+        throw ComAtprotoRepoGetBlob.toKnownErr(e)
       })
   }
