+++ packages/pds/src/api/com/atproto/repo/listRecords.ts
@@ -1,0 +1,38 @@
+import { InvalidRequestError } from '@atproto/xrpc-server'
+import { AtUri } from '@atproto/uri'
+import { Server } from '../../../../lexicon'
+import AppContext from '../../../../context'
+
+export default function (server: Server, ctx: AppContext) {
+  server.com.atproto.repo.listRecords(async ({ params }) => {
+    const { user, collection, limit, before, after, reverse } = params
+
+    const did = await ctx.services.actor(ctx.db).getDidForActor(user)
+    if (!did) {
+      throw new InvalidRequestError(`Could not find user: ${user}`)
+    }
+
+    const records = await ctx.services
+      .record(ctx.db)
+      .listRecordsForCollection(
+        did,
+        collection,
+        limit || 50,
+        reverse || false,
+        before,
+        after,
+      )
+
+    const lastRecord = records.at(-1)
+    const lastUri = lastRecord && new AtUri(lastRecord?.uri)
+
+    return {
+      encoding: 'application/json',
+      body: {
+        records,
+        // Paginate with `before` by default, paginate with `after` when using `reverse`.
+        cursor: lastUri?.rkey,
+      },
+    }
+  })
+}
