+++ packages/pds/src/lexicon/index.ts
@@ -33,6 +33,7 @@ import * as ComAtprotoRepoBatchWrite from './types/com/atproto/repo/batchWrite'
 import * as ComAtprotoRepoCreateRecord from './types/com/atproto/repo/createRecord'
 import * as ComAtprotoRepoDeleteRecord from './types/com/atproto/repo/deleteRecord'
 import * as ComAtprotoRepoDescribe from './types/com/atproto/repo/describe'
+import * as ComAtprotoRepoGetBlob from './types/com/atproto/repo/getBlob'
 import * as ComAtprotoRepoGetRecord from './types/com/atproto/repo/getRecord'
 import * as ComAtprotoRepoListRecords from './types/com/atproto/repo/listRecords'
 import * as ComAtprotoRepoPutRecord from './types/com/atproto/repo/putRecord'
@@ -370,6 +371,13 @@ export class RepoNS {
     cfg: ConfigOf<AV, ComAtprotoRepoDescribe.Handler<ExtractAuth<AV>>>,
   ) {
     const nsid = 'com.atproto.repo.describe' // @ts-ignore
+    return this._server.xrpc.method(nsid, cfg)
+  }
+
+  getBlob<AV extends AuthVerifier>(
+    cfg: ConfigOf<AV, ComAtprotoRepoGetBlob.Handler<ExtractAuth<AV>>>,
+  ) {
+    const nsid = 'com.atproto.repo.getBlob' // @ts-ignore
     return this._server.xrpc.method(nsid, cfg)
   }
